+++ src/components/Header/utils/headerUtils.test.ts
@@ -1,11 +1,11 @@
 import {
-  getEncoreCatalogURL,
+  getCatalogURL,
   getNYPLSearchURL,
   getResearchCatalogURL,
 } from "./headerUtils";
 
 describe("Header utils", () => {
-  describe("getEncoreCatalogURL", () => {
+  describe("getCatalogURL", () => {
     const currentDate = new Date("2022-01-01");
     let realDate;
 
@@ -23,19 +23,19 @@ describe("Header utils", () => {
       (global as any).Date = realDate;
     });
 
-    it("should return a URL for the Encore catalog", () => {
+    it("should return a URL for the Catalog", () => {
       const searchValue = "foo bar";
-      const url = getEncoreCatalogURL(searchValue);
+      const url = getCatalogURL(searchValue);
       expect(url).toEqual(
-        "https://browse.nypl.org/iii/encore/search/C__Sfoo%20bar__Orightresult__U?searched_from=header_search&timestamp=1640995200000&lang=eng"
+        "https://borrow.nypl.org/search?query=foo%20bar&searchType=everything&pageSize=10&searched_from=header_search&timestamp=1640995200000&lang=eng"
       );
     });
 
     it("should return a URL for the Encore catalog with special characters", () => {
       const searchValue = "foo bar/\\?=";
-      const url = getEncoreCatalogURL(searchValue);
+      const url = getCatalogURL(searchValue);
       expect(url).toEqual(
-        "https://browse.nypl.org/iii/encore/search/C__Sfoo%20bar%2F%5C%3F%3D__Orightresult__U?searched_from=header_search&timestamp=1640995200000&lang=eng"
+        "https://borrow.nypl.org/search?query=foo%20bar%2F%5C%3F%3D&searchType=everything&pageSize=10&searched_from=header_search&timestamp=1640995200000&lang=eng"
       );
     });
   });
@@ -58,19 +58,19 @@ describe("Header utils", () => {
       (global as any).Date = realDate;
     });
 
-    it("should return a URL for the Encore catalog", () => {
+    it("should return a URL for the Research Catalog", () => {
       const searchValue = "foo bar";
       const url = getResearchCatalogURL(searchValue);
       expect(url).toEqual(
-        "https://www.nypl.org/research/research-catalog/search?q=foo%20bar&?searched_from=header_search&timestamp=1640995200000&lang=eng"
+        "https://www.nypl.org/research/research-catalog/search?q=foo%20bar&searched_from=header_search&timestamp=1640995200000&lang=eng"
       );
     });
 
-    it("should return a URL for the Encore catalog with special characters", () => {
+    it("should return a URL for the Research Catalog with special characters", () => {
       const searchValue = "foo bar/\\?=";
       const url = getResearchCatalogURL(searchValue);
       expect(url).toEqual(
-        "https://www.nypl.org/research/research-catalog/search?q=foo%20bar%2F%5C%3F%3D&?searched_from=header_search&timestamp=1640995200000&lang=eng"
+        "https://www.nypl.org/research/research-catalog/search?q=foo%20bar%2F%5C%3F%3D&searched_from=header_search&timestamp=1640995200000&lang=eng"
       );
     });
   });